+++ web/app/themes/rebe/src/lib/ReBe/API/Data/Dictionary.php
@@ -32,16 +32,17 @@ class Dictionary
             'timeline_search_result_none_message'    => get_field('timeline_search_result_none_message', 'option'),
 
             // Story post
-            'story_share_story_label'                => get_field('story_share_story_label', 'option'),
-            'story_related_stories_title'            => get_field('story_related_stories_title', 'option'),
-            'story_related_events_title'             => get_field('story_related_events_title', 'option'),
             'story_cc_notice_title'                  => get_field('story_cc_notice_title', 'option'),
             'story_cc_notice_text'                   => get_field('story_cc_notice_text', 'option'),
+            'share_story_label'                      => get_field('share_story_label', 'option'),
+            'story_related_stories_title'            => get_field('story_related_stories_title', 'option'),
+            'story_related_events_title'             => get_field('story_related_events_title', 'option'),
 
             // Event post
             'event_impact_counter_label'             => get_field('event_impact_counter_label', 'option'),
             'event_impact_button_default_label'      => get_field('event_impact_button_default_label', 'option'),
             'event_impact_button_active_label'       => get_field('event_impact_button_active_label', 'option'),
+            'share_event_label'                      => get_field('share_event_label', 'option'),
             'event_related_stories_title'            => get_field('event_related_stories_title', 'option'),
             'event_related_events_title'             => get_field('event_related_events_title', 'option'),
             'event_sources_title'                    => get_field('event_sources_title', 'option'),
@@ -55,12 +56,15 @@ class Dictionary
             'event_add_resource_button_label'        => get_field('event_suggest_resource_button_label', 'option'),
 
             // Common
-            'toggle_navigation_button_label'         => get_field('toggle_navigation_button_screen_text', 'option'),
+            'toggle_navigation_button_screen_text'   => get_field('toggle_navigation_button_screen_text', 'option'),
+            'newsletter_sign_up_button_label'        => get_field('newsletter_sign_up_button_label', 'option'),
             'contact_us_button_label'                => get_field('contact_us_button_label', 'option'),
             'search_label'                           => get_field('search_label', 'option'),
             'close_button_label'                     => get_field('close_button_label', 'option'),
             'next_button_label'                      => get_field('next_button_label', 'option'),
             'previous_button_label'                  => get_field('previous_button_label', 'option'),
+            'follow_label'                           => get_field('follow_label', 'option'),
+            'share_label'                            => get_field('share_label', 'option'),
 
         ];