+++ web/app/themes/rebe/src/lib/ReBe/API/Data/Settings.php
@@ -14,12 +14,19 @@ class Settings
 
         $data = [
 
-            // Basic
-            'site_date_format'    => get_field('site_date_format', 'option'),
+            // Global
             'site_language'       => get_field('site_language', 'option'),
+            'site_date_format'    => get_field('site_date_format', 'option'),
+
+            // Services
             'google_analytics_id' => get_field('google_analytics_id', 'option'),
             'facebook_app_id'     => get_field('facebook_app_id', 'option'),
-            'twitter_handle'      => get_field('twitter_handle', 'option'),
+            'facebook_admin_id'   => get_field('facebook_admin_id', 'option'),
+            'facebook_user_id'    => get_field('facebook_user_id', 'option'),
+            'facebook_user_name'  => get_field('facebook_user_name', 'option'),
+            'twitter_user_name'   => get_field('twitter_user_name', 'option'),
+            'vimeo_user_name'     => get_field('vimeo_user_name', 'option'),
+            'social_hashtags'     => get_field('social_hashtags', 'option'),
 
             // Timeline
             'timelines'           => get_field('public_timelines', 'option'),